requirements.txt
python-dotenv
bs4
pyTelegramBotAPI
